refactor: extract shared interaction error responder and test fixture
Move the acknowledged-aware error response logic out of
SlashCommandHandler into InteractionErrorResponder so other interaction
handlers can reuse it, and promote the recording-proxy fake from
SlashCommandErrorResponseTest to a shared RecordingInteraction test
fixture. No behavior change.
